# backend/api/src/protoBufferTypes/Extra.ts
// Original file: src/protos/order-service.proto


export interface Extra {
  'id'?: (string);
  'name'?: (string);
  'price'?: (number);
}

export interface Extra__Output {
  'id'?: (string);
  'name'?: (string);
  'price'?: (number);
}