# backend/api/src/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ice cream market.",
    "description": "This project it's api to feed the ice cream market application service.",
    "contact": {
      "phone": "+55(84)981758502",
      "email": "haryssonsoares@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "If you run this project locally"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Extra": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/services/orders": {
      "post": {
        "summary": "Create Order",
        "description": "Endpoint to add a new Order to the system.",
        "tags": [
          "Order"
        ],
        "requestBody": {
          "description": "Order details for creation",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "A Test ServiceOrder",
                    "require": false
                  },
                  "size": {
                    "type": "string",
                    "example": "Size test",
                    "require": true
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "received",
                      "making",
                      "delivering",
                      "done"
                    ],
                    "example": "received",
                    "require": true
                  },
                  "paymentMethod": {
                    "type": "string",
                    "enum": [
                      "cash",
                      "pix",
                      "card"
                    ],
                    "example": "card",
                    "require": true
                  },
                  "price": {
                    "type": "number",
                    "example": 9.99,
                    "require": true
                  },
                  "totalPrice": {
                    "type": "number",
                    "example": 99.99,
                    "require": true
                  },
                  "maxCreamsAllowed": {
                    "type": "number",
                    "example": 2,
                    "require": true
                  },
                  "maxToppingsAllowed": {
                    "type": "number",
                    "example": 3,
                    "require": true
                  },
                  "isPaid": {
                    "type": "boolean",
                    "example": true,
                    "require": true
                  },
                  "creams": {
                    "type": "array",
                    "example": [
                      {
                        "id": "0",
                        "name": "Cream test",
                        "price": 9.99
                      },
                      {
                        "id": "1",
                        "name": "Cream test",
                        "price": 9.99
                      }
                    ],
                    "items": {
                      "$ref": "#/components/schemas/CreamToOrder"
                    },
                    "require": true
                  },
                  "toppings": {
                    "type": "array",
                    "example": [
                      {
                        "id": "0",
                        "name": "Topping test",
                        "price": 9.99
                      },
                      {
                        "id": "1",
                        "name": "Topping test",
                        "price": 9.99
                      },
                      {
                        "id": "2",
                        "name": "Topping test",
                        "price": 9.99
                      }
                    ],
                    "items": {
                      "$ref": "#/components/schemas/ToppingToOrder"
                    },
                    "require": false
                  },
                  "extras": {
                    "type": "array",
                    "example": [
                      {
                        "id": "0",
                        "name": "Extra test",
                        "price": 9.99
                      },
                      {
                        "id": "1",
                        "name": "Extra test",
                        "price": 9.99
                      }
                    ],
                    "items": {
                      "$ref": "#/components/schemas/Extra"
                    },
                    "require": false
                  }
                },
                "required": [
                  "price",
                  "paymentMethod",
                  "isPaid",
                  "creams",
                  "maxCreamsAllowed",
                  "maxToppingsAllowed",
                  "size",
                  "totalPrice",
                  "status"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful creating order",
            "content": {
              "application/json": {
                "example": {
                  "message": "Order created successfully",
                  "data": {
                    "id": "65b27738be95fe7b8d0b45c5"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials",
            "content": {
              "application/json": {
                "example": {
                  "message": "Unauthorized: No access token provided"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity - parameters are invalid",
            "content": {
              "application/json": {
                "example": {
                  "message": "isPaid must be a boolean and not empty"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "List Orders",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page to list orders",
            "required": false,
            "schema": {
              "type": "number",
              "default": 1
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "description": "How many orders to return per page",
            "required": false,
            "schema": {
              "type": "number",
              "default": 10
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Order by some field table",
            "required": false,
            "schema": {
              "type": "string",
              "default": "id:asc"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "description": "Filter orders by some fields table",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "name:like:some text here,id:some id here,totalPrice:gt:1000,size:like:some text here"
          }
        ],
        "description": "Retrieve a list of orders based on optional query parameters.",
        "tags": [
          "Order"
        ],
        "responses": {
          "200": {
            "description": "Successful getting order",
            "content": {
              "application/json": {
                "example": {
                  "message": "Orders listed successfully",
                  "data": {
                    "orders": [
                      {
                        "creams": [
                          {
                            "id": "0",
                            "name": "Cream test",
                            "price": 9
                          },
                          {
                            "id": "1",
                            "name": "Cream test",
                            "price": 9
                          }
                        ],
                        "extras": [
                          {
                            "id": "0",
                            "name": "Extra test",
                            "price": 9
                          },
                          {
                            "id": "1",
                            "name": "Extra test",
                            "price": 9
                          }
                        ],
                        "toppings": [
                          {
                            "id": "0",
                            "name": "Topping test",
                            "price": 9
                          },
                          {
                            "id": "1",
                            "name": "Topping test",
                            "price": 9
                          },
                          {
                            "id": "2",
                            "name": "Topping test",
                            "price": 9
                          }
                        ],
                        "id": "65b27738be95fe7b8d0b45c5",
                        "name": "A Test ServiceOrder",
                        "size": "Size test",
                        "paymentMethod": "card",
                        "isPaid": true,
                        "maxCreamsAllowed": 2,
                        "maxToppingsAllowed": 3,
                        "price": 9,
                        "totalPrice": 99,
                        "status": "making",
                        "userId": "5606b121-0371-4649-ace8-0f64f2c35271"
                      },
                      {
                        "creams": [
                          {
                            "id": "0",
                            "name": "Cream test",
                            "price": 9
                          },
                          {
                            "id": "1",
                            "name": "Cream test",
                            "price": 9
                          }
                        ],
                        "extras": [
                          {
                            "id": "0",
                            "name": "Extra test",
                            "price": 9
                          },
                          {
                            "id": "1",
                            "name": "Extra test",
                            "price": 9
                          }
                        ],
                        "toppings": [
                          {
                            "id": "0",
                            "name": "Topping test",
                            "price": 9
                          },
                          {
                            "id": "1",
                            "name": "Topping test",
                            "price": 9
                          },
                          {
                            "id": "2",
                            "name": "Topping test",
                            "price": 9
                          }
                        ],
                        "id": "65b27738be95fe7b8d0b45ce",
                        "name": "B Test ServiceOrder Client",
                        "size": "Size test",
                        "paymentMethod": "card",
                        "isPaid": true,
                        "maxCreamsAllowed": 2,
                        "maxToppingsAllowed": 3,
                        "price": 9,
                        "totalPrice": 99,
                        "status": "received",
                        "userId": "083e8037-185d-45ef-ab02-8521ad957a4c"
                      },
                      {
                        "creams": [
                          {
                            "id": "0",
                            "name": "Cream test",
                            "price": 9
                          },
                          {
                            "id": "1",
                            "name": "Cream test",
                            "price": 9
                          }
                        ],
                        "extras": [
                          {
                            "id": "0",
                            "name": "Extra test",
                            "price": 9
                          },
                          {
                            "id": "1",
                            "name": "Extra test",
                            "price": 9
                          }
                        ],
                        "toppings": [
                          {
                            "id": "0",
                            "name": "Topping test",
                            "price": 9
                          },
                          {
                            "id": "1",
                            "name": "Topping test",
                            "price": 9
                          },
                          {
                            "id": "2",
                            "name": "Topping test",
                            "price": 9
                          }
                        ],
                        "id": "65b27738be95fe7b8d0b45d7",
                        "name": "C Test ServiceOrder Member",
                        "size": "Size test",
                        "paymentMethod": "card",
                        "isPaid": true,
                        "maxCreamsAllowed": 2,
                        "maxToppingsAllowed": 3,
                        "price": 9,
                        "totalPrice": 99,
                        "status": "received",
                        "userId": "11de8153-178b-4913-92e4-dc724258546d"
                      }
                    ],
                    "totalPages": 1
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials",
            "content": {
              "application/json": {
                "example": {
                  "message": "Unauthorized: No access token provided"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity - parameters are invalid",
            "content": {
              "application/json": {
                "example": {
                  "message": "Unknown field(s)"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
...
}